refactor 클래스명 수정 및 불필요한 테스트코드 제거

+++ students/10th/baejunggyoo/js/main.js
@@ -1,50 +1,44 @@
 // 댓글 입력 시 게시 버튼 색 변경 이벤트
-const thisIsReplyForm = document.querySelector('.replyForm');
+const ReplyForm = document.querySelector('.replyForm');
 
-thisIsReplyForm.addEventListener('keyup', function () {
-  const thisIsTextArea = document.querySelector('.reply');
-  const thisIsBtnText = document.querySelector('.btnText');
-  if (thisIsTextArea.value === '') {
-    thisIsBtnText.style.color = '#c4dff6';
-  } else {
-    thisIsBtnText.style.color = '#2e93e2';
-  }
+ReplyForm.addEventListener('keyup', function () {
+  const TextArea = document.querySelector('.reply');
+  const BtnText = document.querySelector('.btnText');
+  TextArea.value === '' ? BtnText.style.color = '#c4dff6' : BtnText.style.color = '#2e93e2';
 });
 
 // 댓글 올리고 삭제하는 이벤트
-const thisIsReplyBtn = document.querySelector('replyUpBtn');
-const thisIsTextArea = document.querySelector('.reply');
+const textArea = document.querySelectorAll('.reply')[0];
 
-thisIsTextArea.addEventListener('keydown', uploadReply);
+textArea.addEventListener('keyup', uploadReply);
 
 function uploadReply(e) {
-  if (e.keyCode === 13) {
+  if (e.keyCode === 13 && textArea.length !== 0) {
     let comment = document.querySelector('.replyContainer');
-    let makeDiv = document.createElement('div');
+    let makeReplyBox = document.createElement('div');
+    let makeUserBox = document.createElement('div');
     let makeUser = document.createElement('span');
     let makeMsg = document.createElement('span');
     let makeBtn = document.createElement('button');
 
-    makeDiv.className = 'replyBox';
+    makeReplyBox.className = 'replyBox';
+    makeUserBox.className = 'userBox';
     makeUser.className = 'replyUser';
     makeMsg.className = 'replyMsg';
-    makeBtn.className = 'deleteBtn';
+    makeBtn.className = 'likeBtn';
 
-    makeMsg.innerHTML = thisIsTextArea.value;
+    makeMsg.innerHTML = textArea.value;
     makeUser.innerHTML = 'juuuuungq';
     makeBtn.innerHTML = '삭제';
 
-    comment.appendChild(makeDiv);
-    makeDiv.appendChild(makeUser);
-    makeDiv.appendChild(makeMsg);
-    makeDiv.appendChild(makeBtn);
-
-    thisIsTextArea.innerHTML = '';
+    comment.appendChild(makeReplyBox);
+    makeReplyBox.appendChild(makeUserBox);
+    makeReplyBox.appendChild(makeBtn);
+    makeUserBox.appendChild(makeUser);
+    makeUserBox.appendChild(makeMsg);
 
     makeBtn.addEventListener('click', (event) => {
       const btn = event.target;
-      console.log(btn);
-
       const deleteAll = btn.parentNode;
       comment.removeChild(deleteAll);
     });
